% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb1992.R
\docType{data}
\name{bb1992}
\alias{bb1992}
\title{Dataset from Babakus & Boller (1992)}
\format{
A data.frame of 22 variables (Q1-Q22) with 502 observations
}
\source{
Provided in paper
}
\usage{
data(bb1992)
}
\description{
Data from Babakus & Boller (1992) who investigated the dimensionality of the SERVQUAL scale based on a sample of N = 502.
The data is available as a data.frame (simulated via mvrnorm in package MASS based on the correlation matrix provided by the authors) and used in the vignette.
}
\examples{
data(bb1992)
head(bb1992, 3)
}
\references{
Babakus, E., & Boller, G. W. (1992). An empirical assessment of the SERVQUAL scale. Journal of Business Research, 24(3), 253–268. https://doi.org/10.1016/0148-2963(92)90022-4
}
\keyword{datasets}
